+++ userpic.php
@@ -15,16 +15,37 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Version information.
+ * Delivery of userpictures to show in collabora as avatar.
  *
  * @package   mod_collabora
  * @copyright 2019 Davo Smith, Synergy Learning
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-defined('MOODLE_INTERNAL') || die;
 
-$plugin->version   = 2025030400;
-$plugin->release   = 'v4.5.3 (2025030400)';
-$plugin->requires  = 2022111800; // Moodle 4.1.
-$plugin->component = 'mod_collabora';
-$plugin->maturity  = MATURITY_BETA;
+use mod_collabora\api\collabora_fs;
+
+// This script is called by the Collabora server and does not need cookies!
+define('NO_MOODLE_COOKIES', true);
+
+require_once(__DIR__ . '/../../config.php');
+require_once($CFG->libdir . '/filelib.php');
+
+$userid = required_param('userid', PARAM_INT);
+$doctoken = required_param('doctoken', PARAM_TEXT);
+
+if (!collabora_fs::validate_doctoken($doctoken, $userid)) {
+    send_file_not_found();
+}
+
+$context = \context_user::instance($userid);
+$fs = get_file_storage();
+if (!$file = $fs->get_file($context->id, 'user', 'icon', 0, '/', 'f1.jpg')) {
+    if (!$file = $fs->get_file($context->id, 'user', 'icon', 0, '/', 'f1.png')) {
+        die;
+    }
+}
+
+if (!empty($file)) {
+    send_stored_file($file);
+}
+send_file_not_found();